Rename LogoutIcon and remove from index
