+++ site/src/components/Icons/index.ts
@@ -1,4 +1,3 @@
 export { CoderIcon } from "./CoderIcon"
 export { Logo } from "./Logo"
-export * from "./Logout"
 export { WorkspacesIcon } from "./WorkspacesIcon"
