+++ site/src/components/Navbar/UserDropdown/index.tsx
@@ -10,7 +10,7 @@ import { Link } from "react-router-dom"
 import { UserResponse } from "../../../api/types"
 import { BorderedMenu } from "../../BorderedMenu/BorderedMenu"
 import { CloseDropdown, OpenDropdown } from "../../DropdownArrows/DropdownArrows"
-import { LogoutIcon } from "../../Icons"
+import { LogoutIcon } from "../../Icons/LogoutIcon"
 import { DocsIcon } from "../../Icons/DocsIcon"
 import { UserAvatar } from "../../User"
 import { UserProfileCard } from "../../User/UserProfileCard"
